+++ src/utils.py
@@ -9,7 +9,7 @@ import numpy as np
 from torchvision import transforms
 from PIL import Image
 
-import sensors
+import metrics
 
 
 
@@ -181,20 +181,20 @@ def build_and_run_device_query():
 
 def get_coral_stats(output_queue, stop_event):
     while not stop_event.is_set():
-        tpu_freq = sensors.handle_coral_dev_board_tpu_freq()[0]
-        cpu_freq = sensors.handle_coral_dev_board_cpu_freq()
-        temp = sensors.handle_coral_dev_board_temp()[0]
-        cpu_usage = sensors.get_cpu_usage()
-        ram_usage = sensors.get_memory_usage()[0]
+        tpu_freq = metrics.handle_coral_dev_board_tpu_freq()[0]
+        cpu_freq = metrics.handle_coral_dev_board_cpu_freq()
+        temp = metrics.handle_coral_dev_board_temp()[0]
+        cpu_usage = metrics.get_cpu_usage()
+        ram_usage = metrics.get_memory_usage()[0]
         output_queue.put((cpu_usage, ram_usage, temp, tpu_freq, cpu_freq))
     
 
 def get_stats_rockpi(output_queue, stop_event):
     while not stop_event.is_set():
-        cpu_freq = sensors.handle_firefly_rk3399_cpu_freq()
-        cpu_usage = sensors.get_cpu_usage()
-        ram_usage = sensors.get_memory_usage()[0]
-        temp = np.sum(sensors.handle_firefly_rk3399_temp())/2
+        cpu_freq = metrics.handle_firefly_rk3399_cpu_freq()
+        cpu_usage = metrics.get_cpu_usage()
+        ram_usage = metrics.get_memory_usage()[0]
+        temp = np.sum(metrics.handle_firefly_rk3399_temp())/2
         output_queue.put((cpu_usage, ram_usage, temp/1000, cpu_freq))
